Introduce discrete distributions PMF chart, add discrete distributions logic, add binomial distribution, refactor dashboard cubit, and do another refactors

+++ lib/feature/distribution_dashboard/domain/use_case/change_distribution_parameter_in_setup_use_case.dart
@@ -16,7 +16,7 @@ class ChangeDistributionParameterInSetupUseCase {
       );
     }
     final newParamsSetup = paramsSetup.copyWith(
-      values: Map.of(paramsSetup.values)..[parameter] = value.toDouble(),
+      values: Map.of(paramsSetup.values)..[parameter] = value,
     );
     await distributionDashboardRepository.setParamsSetup(newParamsSetup);
   }
